Add custom confirmation modal
Added customized confirmation modal that uses ngx-bootstrap library to
create modal that can be used for confirmation operations.

Modified the existing confirmation actions that utilized browser's
default functionality with the customized modal.

Note: the modal library has recently dropped support for bootstrap v3
and therefore, an older compatible library version is used.

+++ src/app/shopping-list/shopping-list-item/shopping-list-item.component.ts
@@ -1,5 +1,10 @@
 import { Component, Input } from '@angular/core';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
+import {
+  ConfirmationResult,
+  ConfirmationType,
+} from '../../shared/models/confirmation.types';
+import { ModalService } from '../../shared/services/modal.service';
 import { ShoppingListItem } from '../models/shopping-list-item-model';
 import { ShoppingListService } from '../services/shopping-list.service';
 
@@ -13,9 +18,32 @@ export class ShoppingListItemComponent {
 
   iconTrash = faTrash;
 
-  constructor(private shoppingListService: ShoppingListService) {}
+  constructor(
+    private shoppingListService: ShoppingListService,
+    private modalService: ModalService
+  ) {}
 
   deleteItem() {
-    this.shoppingListService.deleteItem(this.item!!.ordinal);
+    if (!this.item) {
+      console.debug('delete triggered without item');
+      return;
+    }
+
+    this.modalService.handleConfirmation({
+      confirmationType: ConfirmationType.DELETE,
+      itemDescription: this.item.ingredient.name,
+      onConfirmationResult: (res) => {
+        if (res == ConfirmationResult.YES) {
+          this.shoppingListService.deleteItem(this.item!!.ordinal);
+          console.debug(
+            'deletion completed for "' + this.item?.ingredient.name + '"'
+          );
+        } else {
+          console.debug(
+            'deletion aborted for "' + this.item?.ingredient.name + '"'
+          );
+        }
+      },
+    });
   }
 }
